+++ src/components/feature-board-file-element/content/display/image-display/ImageDisplay.vue
@@ -1,9 +1,18 @@
 <template>
-	<ColorOverlay
-		:isOverlayDisabled="isEditMode"
-		@on:action="openLightBox"
-		color="var(--v-black-base)"
+	<div
+		class="image-display-container"
+		ref="containerRef"
+		:tabindex="tabIndex"
+		@click.stop.prevent="onClick"
+		@focusin.stop.prevent="onFocusIn"
+		@focusout.stop.prevent="onFocusOut"
+		@keydown.enter.stop.prevent="onKeyDown"
+		@keydown.space.stop.prevent="onKeyDown"
+		@mouseenter.stop.prevent="onMouseEnter"
+		@mouseleave.stop.prevent="onMouseLeave"
 	>
+		<div v-if="showOverlay" class="image-display-overlay rounded-t-sm" />
+
 		<img
 			class="image-display-image rounded-t-sm"
 			loading="lazy"
@@ -14,7 +23,7 @@
 		<ContentElementBar class="menu">
 			<template #menu><slot /></template>
 		</ContentElementBar>
-	</ColorOverlay>
+	</div>
 </template>
 
 <script lang="ts">
@@ -22,9 +31,8 @@ import { FileElementResponse } from "@/serverApi/v3";
 import { convertDownloadToPreviewUrl } from "@/utils/fileHelper";
 import { I18N_KEY, injectStrict } from "@/utils/inject";
 import { LightBoxOptions, useLightBox } from "@ui-light-box";
-import { PropType, computed, defineComponent } from "vue";
+import { PropType, computed, defineComponent, ref } from "vue";
 import { ContentElementBar } from "@ui-board";
-import { ColorOverlay } from "@ui-color-overlay";
 
 export default defineComponent({
 	name: "ImageDisplay",
@@ -35,9 +43,14 @@ export default defineComponent({
 		isEditMode: { type: Boolean, required: true },
 		element: { type: Object as PropType<FileElementResponse>, required: true },
 	},
-	components: { ContentElementBar, ColorOverlay },
+	components: { ContentElementBar },
 	setup(props) {
 		const i18n = injectStrict(I18N_KEY);
+		const containerRef = ref<HTMLDivElement | undefined>();
+		const isFocused = ref(false);
+		const isHovered = ref(false);
+
+		const tabIndex = computed(() => (!props.isEditMode ? 0 : -1));
 
 		const alternativeText = computed(() => {
 			const altTranslation = i18n.t(
@@ -49,6 +62,49 @@ export default defineComponent({
 
 			return altText;
 		});
+
+		const showOverlay = computed(
+			() => !props.isEditMode && (isFocused.value || isHovered.value)
+		);
+
+		const onClick = () => {
+			if (!props.isEditMode) {
+				openLightBox();
+			}
+			if (!props.isEditMode) {
+				containerRef.value?.blur();
+			}
+		};
+
+		const onFocusIn = () => {
+			if (!props.isEditMode) {
+				isFocused.value = true;
+			}
+		};
+
+		const onFocusOut = () => {
+			if (!props.isEditMode) {
+				isFocused.value = false;
+			}
+		};
+
+		const onKeyDown = () => {
+			if (!props.isEditMode) {
+				openLightBox();
+			}
+		};
+
+		const onMouseEnter = () => {
+			if (!props.isEditMode) {
+				isHovered.value = true;
+			}
+		};
+
+		const onMouseLeave = () => {
+			if (!props.isEditMode) {
+				isHovered.value = false;
+			}
+		};
 
 		const openLightBox = () => {
 			const previewUrl = convertDownloadToPreviewUrl(props.src);
@@ -67,13 +123,38 @@ export default defineComponent({
 
 		return {
 			alternativeText,
-			openLightBox,
+			containerRef,
+			showOverlay,
+			tabIndex,
+			onClick,
+			onFocusIn,
+			onFocusOut,
+			onKeyDown,
+			onMouseEnter,
+			onMouseLeave,
 		};
 	},
 });
 </script>
 
 <style scoped>
+.image-display-container {
+	position: relative;
+	min-height: 52px;
+	display: flex;
+	align-items: center;
+}
+
+.image-display-overlay {
+	position: absolute;
+	top: 0;
+	left: 0;
+	width: 100%;
+	height: 100%;
+	background: var(--v-black-base);
+	opacity: 0.2;
+}
+
 .image-display-image {
 	pointer-events: none;
 	display: block;
