--- src/components/ui-color-overlay/index.ts
@@ -1,3 +1,0 @@
-import ColorOverlay from "./ColorOverlay.vue";
-
-export { ColorOverlay };
